separating logger name from realm
So far, Realms were used to enrich the logger name.

This collides with the idea of absolute realm names, bcause
logger names are additive, the can just be extended but not replaced.

Therefore, we introduce a new message context type *Name*, which
is used to enrich the the logger name, only.

Similar to *Tag*, which is used to control the log state, only, *Name*
is used, to enrich the logger name, only.

The Realm is now mapped to logr as additional special log
message attribute with the name `realm`.

ATTENTION: this an incompatible change for the resulting log entries.
The realm is not part of the logger name, anymore.